refactor(gtfs-transform): move code into smaller files
