+++ lib/manager/components/FeedSourceSettings.js
@@ -13,7 +13,7 @@ import {LinkContainer} from 'react-router-bootstrap'
 import * as feedsActions from '../actions/feeds'
 import toSentenceCase from '../../common/util/to-sentence-case'
 import ExternalPropertiesTable from './ExternalPropertiesTable'
-import FeedTransformationSettings from './FeedTransformationSettings'
+import FeedTransformationSettings from './transform/FeedTransformationSettings'
 import GeneralSettings from './GeneralSettings'
 
 import type {Feed, Project} from '../../types'
@@ -98,14 +98,13 @@ export default class FeedSourceSettings extends Component<Props> {
             user={user}
           />
           : showTransformationsTab
-            // FIXME: Remove props spread.
             ? <FeedTransformationSettings
               disabled={disabled}
               feedSource={feedSource}
               project={project}
               updateFeedSource={this.props.updateFeedSource}
               user={user}
-             />
+            />
             : <Col xs={7}>
               <ExternalPropertiesTable
                 resourceType={resourceType}
