+++ lib/types/index.js
@@ -293,12 +293,13 @@ export type RetrievalMethod =
   'VERSION_CLONE'
 
 export type FeedTransformation = {
+  '@type': 'DeleteRecordsTransformation' | 'ReplaceFileFromVersionTransformation' | 'ReplaceFileFromStringTransformation',
   active: boolean,
   csvData?: string,
   matchField?: string,
   matchValues?: Array<string>,
   sourceVersionId?: string,
-  type: 'DeleteRecordsTransformation' | 'ReplaceFileFromVersionTransformation' | 'ReplaceFileFromStringTransformation',
+  table: string
 }
 
 export type FeedTransformRules = {
